requirements.txt
aiobotocore @ file:///tmp/build/80754af9/aiobotocore_1613245695822/work
aiohttp @ file:///opt/concourse/worker/volumes/live/c42dc273-cfa8-45cb-6a1f-5108aef22543/volume/aiohttp_1607109648996/work
aioitertools @ file:///tmp/build/80754af9/aioitertools_1607109665762/work
alpaca-trade-api==0.53.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
APMonitor==0.34
appdirs==1.4.4
appnope==0.1.0
async-timeout==3.0.1
asyncpg==0.22.0
attrs==19.3.0
awscli==1.19.9
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bleach==3.1.5
boto3==1.17.9
botocore==1.20.9
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clyent==1.2.2
colorama==0.4.3
conda==4.9.2
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
control==0.8.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discrete==0.0.2
distlib==0.3.1
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
fsspec @ file:///tmp/build/80754af9/fsspec_1602684995936/work
future==0.18.2
gekko==0.2.8
glob2==0.7
gunicorn==20.0.4
gurobipy==9.0.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///opt/concourse/worker/volumes/live/84197498-cbc0-4436-7ce0-03c4490b7a28/volume/importlib-metadata_1593446431408/work
imutils==0.5.3
iniconfig==1.1.1
ipykernel @ file:///opt/concourse/worker/volumes/live/18b922c7-3041-4126-5434-cac748d45c19/volume/ipykernel_1594755374448/work/dist/ipykernel-5.3.2-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/7ea70a71-5624-4799-50b3-3b90e07922f3/volume/ipython_1593447385860/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi @ file:///opt/concourse/worker/volumes/live/114a62c0-dd60-4a40-6382-85abb31b5f6e/volume/jedi_1592841953086/work
jellyfish==0.8.2
Jinja2==2.11.2
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
kiwisolver==1.2.0
lib50==2.0.9
libarchive-c==2.9
llvmlite==0.33.0
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mpmath==1.1.0
multidict @ file:///opt/concourse/worker/volumes/live/36866b75-3506-404d-4bf8-25adbcd7e41f/volume/multidict_1600456436970/work
multitasking==0.0.9
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numba==0.50.1
numpy==1.19.0
olefile==0.46
opencv-contrib-python==4.4.0.46
opencv-python-headless==4.4.0.46
packaging==20.4
pandas==1.2.1
pandas-datareader==0.9.0
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///opt/concourse/worker/volumes/live/05b3bc99-cf9f-44f2-42cf-267ed3a55922/volume/pillow_1594307311725/work
pip-autoremove==0.9.1
pkginfo==1.5.0.1
pluggy==0.13.1
praw==7.1.4
prawcore==1.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psaw==0.1.0
psutil==5.7.0
psycopg2 @ file:///opt/concourse/worker/volumes/live/ad0d2468-72e7-4502-526f-2fd0ec5d33ff/volume/psycopg2_1612298588979/work
ptyprocess==0.6.0
py==1.10.0
py4j==0.10.9
pyasn1==0.4.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pygame==2.0.1
Pygments==2.6.1
pymongo==3.11.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pyspark==3.0.1
pytest==6.2.2
pytest-dotenv==0.5.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
quantecon==0.4.8
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
rsa==4.5
ruamel-yaml==0.15.87
s3fs @ file:///tmp/build/80754af9/s3fs_1608058389258/work
s3transfer @ file:///tmp/build/80754af9/s3transfer_1610404175646/work
scikit-learn==0.24.1
scipy==1.5.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.23
submit50==3.0.3
swagger-client==1.0.0
sympy==1.6.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1593446365756/work
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
update-checker==0.18.0
urllib3==1.25.9
vaderSentiment==3.3.2
virtualenv==20.4.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xmltodict==0.12.0
yarl @ file:///opt/concourse/worker/volumes/live/33e35c6f-e0f3-4e8f-41a6-47322106cfc8/volume/yarl_1598045184895/work
yfinance==0.1.55
zipp==3.1.0
